+++ streamlit_canary/__init__.py
@@ -1,7 +1,9 @@
+# from . import components
+from . import opener
 from . import session
 from .components import *
 from .compositor import Compositor
-from .event_handler_no_rerun import evt
+from .event_loop import event_loop
 from .flow import post_events
 from .opener import open_file
 from .opener import open_folder
@@ -13,6 +15,5 @@ from .session import init_shared_data
 from .session import init_state
 from .session import init_state as get_state
 from .session import shared_data
-# from .wrapper import *
 
-__version__ = '0.2.1'
+__version__ = '0.3.0'