+++ streamlit_canary/components/__init__.py
@@ -1,7 +1,6 @@
 from ._typing import AnyComponent
 from ._typing import AnyContainer
 from .button import long_button
-from .card import bordered_container
 from .card import card
 from .column import columns
 from .filelist import filelist
@@ -12,4 +11,8 @@ from .progress import ProgressItem
 from .progress import progress
 from .radio import radio
 from .row import row
-from .text import hint
+from .toggle import toggled_button
+from .tree_select import ask_files
+from .tree_select import ask_folders
+from .tree_select import single_select_dialog
+from .tree_select import tree_select