+++ streamlit_canary/components/card.py
@@ -1,14 +1,10 @@
+import streamlit as st
 import typing as t
 from contextlib import contextmanager
-from functools import partial
-
-import streamlit as st
-
-bordered_container = partial(st.container, border=True)
 
 
 @contextmanager
-def card(title: str = None) -> t.Iterator:
+def card(title: str = '') -> t.Generator:
     with st.container(border=True):
         if title:
             st.write(':blue[**{}**]'.format(title))