+++ streamlit_canary/duplicate_key_resolver.py
@@ -1,0 +1,6 @@
+class UniqueKeyGenerator:
+    def __init__(self, prefix: str) -> None:
+        self._prefix = prefix
+
+    def __call__(self, *args) -> str:
+        return '{}:{}'.format(self._prefix, ':'.join(args))